+++ web/ui/src/features/component/ComponentList.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import { HealthLabel } from './HealthLabel';
-import { ComponentInfo } from './types';
+import { HealthLabel } from '../component/HealthLabel';
+import { ComponentInfo } from '../component/types';
+import Table from '../widget/Table';
 
 import styles from './ComponentList.module.css';
 
@@ -11,29 +12,27 @@ interface ComponentListProps {
 }
 
 const ComponentList = ({ components }: ComponentListProps) => {
+  const tableHeaders = ['Health', 'ID'];
+
+  const renderTableData = () => {
+    return components.map(({ health, id }) => (
+      <tr key={id}>
+        <td>
+          <HealthLabel health={health.state} />
+        </td>
+        <td>
+          {id}
+          <NavLink to={'/component/' + id} className={styles.viewButton}>
+            View
+          </NavLink>
+        </td>
+      </tr>
+    ));
+  };
+
   return (
     <div className={styles.list}>
-      <table className={styles.table}>
-        <tr>
-          <th>Health</th>
-          <th>ID</th>
-        </tr>
-        {components.map((component) => {
-          return (
-            <tr>
-              <td>
-                <HealthLabel health={component.health.state} />
-              </td>
-              <td>
-                {component.id}
-                <NavLink to={'/component/' + component.id} className={styles.viewButton}>
-                  View
-                </NavLink>
-              </td>
-            </tr>
-          );
-        })}
-      </table>
+      <Table tableHeaders={tableHeaders} renderTableData={renderTableData} />
     </div>
   );
 };
