+++ web/ui/src/features/river-js/RiverValue.module.css
@@ -12,6 +12,8 @@
 
 .value span.string {
   color: green;
+  width: 100%;
+  word-wrap: break-word;
 }
 
 .value span.special {
