+++ web/ui/src/features/river-js/RiverValue.tsx
@@ -1,4 +1,5 @@
-import { FC, Fragment, ReactElement } from 'react';
+import React, { FC, Fragment, ReactElement } from 'react';
+import { greatestIndex } from 'd3';
 
 import { ObjectField, Value, ValueType } from './types';
